requirements.txt
opencv-python-headless
mediapipe
scikit-learn
streamlit
numpy
pillow